valheim_infra/requirements.txt
aws-cdk-lib==2.24.1
constructs>=10.0.0,<11.0.0
boto3==1.23.5
httpx==0.22.0
